Crate makefile

This crate has no code, only rustdoc.

Makefile macros

Since we use make as the top level CLI/CI, it would be convenient to wrap Rust nigthly toolchain commands with Makefile macros.

RUSTC = rustup run nightly rustc
RUSTDOC = rustup run nightly rustdoc --out-dir docs

Here is how to use it.

make versions

$(RUSTC) --version
$(RUSTDOC) --version

% make versions
rustup run nightly rustc --version
rustc 1.95.0-nightly (eda76d9d1 2026-01-21)
rustup run nightly rustdoc --out-dir docs  --version
rustdoc 1.95.0-nightly (eda76d9d1 2026-01-21)

make build

$(RUSTDOC) src/introduction.rs
$(RUSTDOC) src/assert_matches.rs
$(RUSTDOC) src/bin/coroutine.rs
$(RUSTDOC) src/makefile.rs
$(RUSTDOC) src/workflow.rs

% make build
rustup run nightly rustdoc --out-dir docs  src/introduction.rs
rustup run nightly rustdoc --out-dir docs  src/assert_matches.rs
rustup run nightly rustdoc --out-dir docs  src/bin/coroutine.rs
rustup run nightly rustdoc --out-dir docs  src/makefile.rs
rustup run nightly rustdoc --out-dir docs  src/workflow.rs

make test

    $(RUSTC) --test src/assert_matches.rs && ./assert_matches && rm ./assert_matches
    $(RUSTC) src/bin/coroutine.rs && ./coroutine && rm ./coroutine

% make test
rustup run nightly rustc --test src/assert_matches.rs && ./assert_matches && rm ./assert_matches

running 1 test
test tests::test_assert_matches - should panic ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.00s

rustup run nightly rustc src/bin/coroutine.rs && ./coroutine && rm ./coroutine
It works!
```